refactor: split monolithic App.js into components/constants/utils
