+++ src/constants/icons.js
@@ -1,0 +1,21 @@
+import {
+  Globe, Download, FileText, Calendar, Video, Music, ShoppingCart,
+  Link as LinkIcon, Youtube, Facebook, MessageCircle
+} from 'lucide-react';
+
+// --- ICONS MAPPING ---
+const ICON_MAP = {
+  link: LinkIcon,
+  download: Download,
+  file: FileText,
+  calendar: Calendar,
+  video: Video,
+  music: Music,
+  shop: ShoppingCart,
+  youtube: Youtube,
+  facebook: Facebook,
+  whatsapp: MessageCircle,
+  globe: Globe
+};
+
+export { ICON_MAP };
